Moved stuff under tracking into its own folder, further splitted into subfolders for classes
